Moved LoadLookupData from Process to Data.
