+++ CHRISUpdate/Validation/ValidateHR.cs
@@ -1,9 +1,9 @@
 ﻿using FluentValidation;
 using FluentValidation.Results;
+using HRUpdate.Data;
 using HRUpdate.Lookups;
 using HRUpdate.Mapping;
 using HRUpdate.Models;
-using HRUpdate.Process;
 using System.Collections.Generic;
 using System.Linq;
